made all the requested changes
