+++ src/modules/client/client/services/client.service.ts
@@ -16,6 +16,9 @@ import { RoleClientService } from '@modules/authorization/services/role-client.s
 import { TeamMember } from '../dtos/team-member.dto';
 import { throwIfNotFound } from '@src/utils/exceptions/common.exception';
 import { ResetClientPasswordDto } from '@modules/authentication/dtos/admin-reset-client-password.dto';
+import { RoleEnum } from '@src/utils/enums/role.enums';
+import { RegisterClientDto } from '../dtos/register-client.dto';
+import { RoleExternalService } from '@modules/authorization/services/role-external.service';
 
 
 @Injectable()
@@ -24,6 +27,7 @@ export class ClientService {
     private clientRepository: ClientRepository,
     private roleClientService: RoleClientService,
     private bcryptEncryptionService: BcryptEncryptionService,
+    private readonly roleExternalService:RoleExternalService
     
   ) {}
 
@@ -184,6 +188,38 @@ export class ClientService {
     await this.clientRepository.getRepository().save(client);
 
     return 'Password successfully reset';
+  }
+
+  async createClient(clientDto: RegisterClientDto): Promise<ClientEntity> {
+    // Check if a client with the provided email already exists
+    const getClientWithProvidedEmail = await this.getClientByEmailOrThrow(clientDto.email);
+  
+    // Check if a client with the provided contact number already exists
+    const getClientWithContactNumber = await this.getClientByContactNumber(clientDto.contactNumber);
+  
+    // Throw an exception if either the email or contact number already exists
+    if (getClientWithContactNumber || getClientWithProvidedEmail) {
+      throw new BadRequestException('User already exists with the provided details');
+    }
+  
+    // Encrypt the password if it is provided, or use a default encrypted password (e.g., encrypted name)
+    const encryptedPassword = clientDto.password 
+      ? await this.bcryptEncryptionService.encrypt(clientDto.password)
+      : await this.bcryptEncryptionService.encrypt(clientDto.name);
+  
+    // Retrieve the default role for clients
+    const role = await this.roleExternalService.getRoleDetails(RoleEnum.OWNER);
+    const roleId = role.id;
+  
+    // Prepare the client data to be saved
+    const clientDataToSave = {
+      ...clientDto,
+      password: encryptedPassword,
+      roleId,
+    };
+  
+    // Save the client data in the database
+    return this.clientRepository.getRepository().save(clientDataToSave);
   }
